[Website] Split Playground remote initialization logic into Blueprint-version specific workers
